refactor: App.tsx를 쉘 + 라우트별 화면으로 분리
- src/screens/{Checklist,Events,Settings}Screen 추출, App은 라우트 분기·사이드바·하단 네비·인증·체크 동기화만 담당
- 하단 네비는 쉘에 한 인스턴스로 유지해 화면 전환 시 포커스 보존 (테스트 추가)
- useAppRoute: navigate 후 도착하는 hashchange가 같은 라우트면 무시 → 설정→검색 진입 시 시트가 닫히던 버그 수정 (테스트 추가)
- 검색/필터 상태를 useChecklistFilters로 묶음
- 미사용 useDetailRoute / parseDetailId / detailHash 삭제, 숨김 처리된 설정 버튼 dead code 삭제
- READ_STALE_TIME → lib/query.ts
- 타이밍 의존이던 #45 병합 실패 테스트는 waitFor로 대기
