+++ src/lib/query.ts
@@ -1,0 +1,2 @@
+/** 읽기 전용 데이터(events/circles) 캐시 신선도. */
+export const READ_STALE_TIME = 5 * 60 * 1000;
